refactor(adk): extract model timeout middleware
