+++ adk/wrappers.go
@@ -44,7 +44,6 @@ type typedModelWrapperConfig[M MessageType] struct {
 	middlewares    []AgentMiddleware
 	retryConfig    *TypedModelRetryConfig[M]
 	failoverConfig *ModelFailoverConfig[M]
-	timeoutConfig  *ModelTimeoutConfig
 	toolInfos      []*schema.ToolInfo
 	cancelContext  *cancelContext
 }
@@ -74,7 +73,6 @@ func buildModelWrappersImpl[M MessageType](m model.BaseModel[M], config *typedMo
 		toolInfos:           config.toolInfos,
 		modelRetryConfig:    config.retryConfig,
 		modelFailoverConfig: config.failoverConfig,
-		modelTimeoutConfig:  config.timeoutConfig,
 		cancelContext:       config.cancelContext,
 	}
 
@@ -288,18 +286,13 @@ func (w *typedEventSenderModelWrapper[M]) WrapModel(_ context.Context, m model.B
 	if mc != nil {
 		failoverConfig = mc.ModelFailoverConfig
 	}
-	var timeoutConfig *ModelTimeoutConfig
-	if mc != nil {
-		timeoutConfig = mc.ModelTimeoutConfig
-	}
-	return &typedEventSenderModel[M]{inner: inner, modelRetryConfig: retryConfig, modelFailoverConfig: failoverConfig, modelTimeoutConfig: timeoutConfig}, nil
+	return &typedEventSenderModel[M]{inner: inner, modelRetryConfig: retryConfig, modelFailoverConfig: failoverConfig}, nil
 }
 
 type typedEventSenderModel[M MessageType] struct {
 	inner               model.BaseModel[M]
 	modelRetryConfig    *TypedModelRetryConfig[M]
 	modelFailoverConfig *ModelFailoverConfig[M]
-	modelTimeoutConfig  *ModelTimeoutConfig
 }
 
 func sendSessionTimelineEvent[M MessageType](ctx context.Context, se *SessionEvent[M]) {
@@ -418,12 +411,16 @@ func modelSpanCompletionMeta[M MessageType](ctx context.Context, startEventID st
 	meta.Usage = modelUsageFromAssistant(msg)
 	meta.FinishReason = assistantFinishReason(msg)
 	meta.Accepted = accepted
-	if timeoutErr, ok := AsModelTimeout(err); ok {
+	var timeoutErr interface {
+		ModelTimeoutSpanMeta() (phase string, timeout time.Duration, elapsed time.Duration, chunksReceived int)
+	}
+	if errors.As(err, &timeoutErr) {
+		phase, timeout, elapsed, chunksReceived := timeoutErr.ModelTimeoutSpanMeta()
 		meta.Timeout = &ModelTimeoutMeta{
-			Phase:          string(timeoutErr.Phase),
-			TimeoutMS:      timeoutErr.Timeout.Milliseconds(),
-			ElapsedMS:      timeoutErr.Elapsed.Milliseconds(),
-			ChunksReceived: timeoutErr.ChunksReceived,
+			Phase:          phase,
+			TimeoutMS:      timeout.Milliseconds(),
+			ElapsedMS:      elapsed.Milliseconds(),
+			ChunksReceived: chunksReceived,
 		}
 	}
 	return meta
@@ -1725,7 +1722,6 @@ type typedStateModelWrapper[M MessageType] struct {
 	toolInfos           []*schema.ToolInfo
 	modelRetryConfig    *TypedModelRetryConfig[M]
 	modelFailoverConfig *ModelFailoverConfig[M]
-	modelTimeoutConfig  *ModelTimeoutConfig
 	cancelContext       *cancelContext
 }
 
@@ -1773,7 +1769,6 @@ func (w *typedStateModelWrapper[M]) wrapGenerateEndpoint(endpoint typedGenerateE
 	hasUserEventSender := w.hasUserEventSender()
 	retryConfig := w.modelRetryConfig
 	failoverConfig := w.modelFailoverConfig
-	timeoutConfig := w.modelTimeoutConfig
 	cc := w.cancelContext
 
 	for i := len(w.handlers) - 1; i >= 0; i-- {
@@ -1783,21 +1778,13 @@ func (w *typedStateModelWrapper[M]) wrapGenerateEndpoint(endpoint typedGenerateE
 		endpoint = func(ctx context.Context, input []M, opts ...model.Option) (M, error) {
 			baseOpts := &model.Options{Tools: baseToolInfos}
 			commonOpts := model.GetCommonOptions(baseOpts, opts...)
-			mc := &TypedModelContext[M]{Tools: commonOpts.Tools, ModelRetryConfig: retryConfig, ModelTimeoutConfig: timeoutConfig, cancelContext: cc}
+			mc := &TypedModelContext[M]{Tools: commonOpts.Tools, ModelRetryConfig: retryConfig, cancelContext: cc}
 			wrappedModel, err := handler.WrapModel(ctx, &typedEndpointModel[M]{generate: innerEndpoint}, mc)
 			if err != nil {
 				var zero M
 				return zero, err
 			}
 			return wrappedModel.Generate(ctx, input, opts...)
-		}
-	}
-
-	if isModelTimeoutConfigActive(timeoutConfig) {
-		innerEndpoint := endpoint
-		endpoint = func(ctx context.Context, input []M, opts ...model.Option) (M, error) {
-			timeoutWrapper := newTypedTimeoutModelWrapper[M](&typedEndpointModel[M]{generate: innerEndpoint}, timeoutConfig)
-			return timeoutWrapper.Generate(ctx, input, opts...)
 		}
 	}
 
@@ -1811,7 +1798,7 @@ func (w *typedStateModelWrapper[M]) wrapGenerateEndpoint(endpoint typedGenerateE
 			if execCtx == nil || execCtx.generator == nil {
 				return innerEndpoint(ctx, input, opts...)
 			}
-			mc := &TypedModelContext[M]{ModelRetryConfig: retryConfig, ModelFailoverConfig: failoverConfig, ModelTimeoutConfig: timeoutConfig, cancelContext: cc}
+			mc := &TypedModelContext[M]{ModelRetryConfig: retryConfig, ModelFailoverConfig: failoverConfig, cancelContext: cc}
 			wrappedModel, err := eventSender.WrapModel(ctx, &typedEndpointModel[M]{generate: innerEndpoint}, mc)
 			if err != nil {
 				var zero M
@@ -1870,7 +1857,6 @@ func (w *typedStateModelWrapper[M]) wrapStreamEndpoint(endpoint typedStreamEndpo
 	hasUserEventSender := w.hasUserEventSender()
 	retryConfig := w.modelRetryConfig
 	failoverConfig := w.modelFailoverConfig
-	timeoutConfig := w.modelTimeoutConfig
 	cc := w.cancelContext
 
 	for i := len(w.handlers) - 1; i >= 0; i-- {
@@ -1880,20 +1866,12 @@ func (w *typedStateModelWrapper[M]) wrapStreamEndpoint(endpoint typedStreamEndpo
 		endpoint = func(ctx context.Context, input []M, opts ...model.Option) (*schema.StreamReader[M], error) {
 			baseOpts := &model.Options{Tools: baseToolInfos}
 			commonOpts := model.GetCommonOptions(baseOpts, opts...)
-			mc := &TypedModelContext[M]{Tools: commonOpts.Tools, ModelRetryConfig: retryConfig, ModelTimeoutConfig: timeoutConfig, cancelContext: cc}
+			mc := &TypedModelContext[M]{Tools: commonOpts.Tools, ModelRetryConfig: retryConfig, cancelContext: cc}
 			wrappedModel, err := handler.WrapModel(ctx, &typedEndpointModel[M]{stream: innerEndpoint}, mc)
 			if err != nil {
 				return nil, err
 			}
 			return wrappedModel.Stream(ctx, input, opts...)
-		}
-	}
-
-	if isModelTimeoutConfigActive(timeoutConfig) {
-		innerEndpoint := endpoint
-		endpoint = func(ctx context.Context, input []M, opts ...model.Option) (*schema.StreamReader[M], error) {
-			timeoutWrapper := newTypedTimeoutModelWrapper[M](&typedEndpointModel[M]{stream: innerEndpoint}, timeoutConfig)
-			return timeoutWrapper.Stream(ctx, input, opts...)
 		}
 	}
 
@@ -1907,7 +1885,7 @@ func (w *typedStateModelWrapper[M]) wrapStreamEndpoint(endpoint typedStreamEndpo
 			if execCtx == nil || execCtx.generator == nil {
 				return innerEndpoint(ctx, input, opts...)
 			}
-			mc := &TypedModelContext[M]{ModelRetryConfig: retryConfig, ModelFailoverConfig: failoverConfig, ModelTimeoutConfig: timeoutConfig, cancelContext: cc}
+			mc := &TypedModelContext[M]{ModelRetryConfig: retryConfig, ModelFailoverConfig: failoverConfig, cancelContext: cc}
 			wrappedModel, err := eventSender.WrapModel(ctx, &typedEndpointModel[M]{stream: innerEndpoint}, mc)
 			if err != nil {
 				return nil, err
@@ -1983,7 +1961,7 @@ func (w *typedStateModelWrapper[M]) Generate(ctx context.Context, _ []M, opts ..
 
 	baseOpts := &model.Options{Tools: w.toolInfos}
 	commonOpts := model.GetCommonOptions(baseOpts, opts...)
-	mc := &TypedModelContext[M]{Tools: commonOpts.Tools, ModelRetryConfig: w.modelRetryConfig, ModelTimeoutConfig: w.modelTimeoutConfig, cancelContext: w.cancelContext}
+	mc := &TypedModelContext[M]{Tools: commonOpts.Tools, ModelRetryConfig: w.modelRetryConfig, cancelContext: w.cancelContext}
 	for _, handler := range w.handlers {
 		var err error
 		ctx, state, err = handler.BeforeModelRewriteState(ctx, state, mc)
@@ -2111,7 +2089,7 @@ func (w *typedStateModelWrapper[M]) Stream(ctx context.Context, _ []M, opts ...m
 
 	baseOpts := &model.Options{Tools: w.toolInfos}
 	commonOpts := model.GetCommonOptions(baseOpts, opts...)
-	mc := &TypedModelContext[M]{Tools: commonOpts.Tools, ModelRetryConfig: w.modelRetryConfig, ModelTimeoutConfig: w.modelTimeoutConfig, cancelContext: w.cancelContext}
+	mc := &TypedModelContext[M]{Tools: commonOpts.Tools, ModelRetryConfig: w.modelRetryConfig, cancelContext: w.cancelContext}
 	for _, handler := range w.handlers {
 		var err error
 		ctx, state, err = handler.BeforeModelRewriteState(ctx, state, mc)
