+++ adk/wrappers_test.go
@@ -17,11 +17,15 @@
 package adk
 
 import (
+	"bytes"
 	"context"
+	"encoding/gob"
 	"errors"
+	"fmt"
 	"sync"
 	"sync/atomic"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -2183,4 +2187,1412 @@ func TestExtractToolIdentifiersToolSearchResult(t *testing.T) {
 	toolName, callID := extractToolIdentifiers(msg)
 	assert.Equal(t, "tool_search", toolName)
 	assert.Equal(t, "call_1", callID)
+}
+
+func TestBuildModelWrappers_FailoverProxyInner(t *testing.T) {
+	base := &fakeChatModel{
+		callbacksEnabled: true,
+		generate: func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.Message, error) {
+			return schema.AssistantMessage("ok", nil), nil
+		},
+		stream: func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.StreamReader[*schema.Message], error) {
+			return schema.StreamReaderFromArray([]*schema.Message{schema.AssistantMessage("ok", nil)}), nil
+		},
+	}
+
+	failoverCfg := &ModelFailoverConfig[*schema.Message]{
+		MaxRetries:     0,
+		ShouldFailover: func(context.Context, *schema.Message, error) bool { return false },
+		GetFailoverModel: func(_ context.Context, _ *FailoverContext[*schema.Message]) (model.BaseChatModel, []*schema.Message, error) {
+			return base, nil, nil
+		},
+	}
+
+	wrapped := buildModelWrappers[*schema.Message](base, &modelWrapperConfig{
+		failoverConfig: failoverCfg,
+	})
+
+	smw, ok := wrapped.(*stateModelWrapper)
+	require.True(t, ok)
+	_, ok = smw.inner.(*failoverProxyModel)
+	require.True(t, ok)
+	require.Same(t, base, smw.original)
+	require.Same(t, failoverCfg, smw.modelFailoverConfig)
+}
+
+func TestStateModelWrapper_Generate_WithFailover(t *testing.T) {
+	wantErr := errors.New("first failed")
+	var shouldCalls int32
+	var m1Calls int32
+	var m2Calls int32
+
+	m1 := &fakeChatModel{
+		callbacksEnabled: true,
+		generate: func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.Message, error) {
+			atomic.AddInt32(&m1Calls, 1)
+			return schema.AssistantMessage("partial", nil), wantErr
+		},
+		stream: func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.StreamReader[*schema.Message], error) {
+			return nil, errors.New("unused")
+		},
+	}
+	m2 := &fakeChatModel{
+		callbacksEnabled: true,
+		generate: func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.Message, error) {
+			atomic.AddInt32(&m2Calls, 1)
+			return schema.AssistantMessage("ok", nil), nil
+		},
+		stream: func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.StreamReader[*schema.Message], error) {
+			return nil, errors.New("unused")
+		},
+	}
+
+	failoverCfg := &ModelFailoverConfig[*schema.Message]{
+		MaxRetries: 1,
+		ShouldFailover: func(_ context.Context, out *schema.Message, err error) bool {
+			atomic.AddInt32(&shouldCalls, 1)
+			require.ErrorIs(t, err, wantErr)
+			require.NotNil(t, out)
+			require.Equal(t, "partial", out.Content)
+			return true
+		},
+		GetFailoverModel: func(_ context.Context, failoverCtx *FailoverContext[*schema.Message]) (model.BaseChatModel, []*schema.Message, error) {
+			require.Equal(t, uint(1), failoverCtx.FailoverAttempt)
+			return m2, nil, nil
+		},
+	}
+
+	wrapped := buildModelWrappers[*schema.Message](m1, &modelWrapperConfig{
+		failoverConfig: failoverCfg,
+	})
+
+	ctx := withTypedChatModelAgentExecCtx(context.Background(), &chatModelAgentExecCtx{
+		failoverLastSuccessModel: m1,
+	})
+	got, err := wrapped.Generate(ctx, []*schema.Message{schema.UserMessage("hi")})
+	require.NoError(t, err)
+	require.NotNil(t, got)
+	require.Equal(t, "ok", got.Content)
+	require.Equal(t, int32(1), atomic.LoadInt32(&m1Calls))
+	require.Equal(t, int32(1), atomic.LoadInt32(&m2Calls))
+	require.Equal(t, int32(1), atomic.LoadInt32(&shouldCalls))
+}
+
+func TestStateModelWrapper_Stream_WithFailover(t *testing.T) {
+	streamErr := errors.New("mid error")
+	var shouldCalls int32
+	var m1Calls int32
+	var m2Calls int32
+
+	m1 := &fakeChatModel{
+		callbacksEnabled: true,
+		generate: func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.Message, error) {
+			return nil, errors.New("unused")
+		},
+		stream: func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.StreamReader[*schema.Message], error) {
+			atomic.AddInt32(&m1Calls, 1)
+			return streamWithMidError([]*schema.Message{
+				schema.AssistantMessage("p1", nil),
+				schema.AssistantMessage("p2", nil),
+			}, streamErr), nil
+		},
+	}
+	m2 := &fakeChatModel{
+		callbacksEnabled: true,
+		generate: func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.Message, error) {
+			return nil, errors.New("unused")
+		},
+		stream: func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.StreamReader[*schema.Message], error) {
+			atomic.AddInt32(&m2Calls, 1)
+			return schema.StreamReaderFromArray([]*schema.Message{schema.AssistantMessage("final", nil)}), nil
+		},
+	}
+
+	failoverCfg := &ModelFailoverConfig[*schema.Message]{
+		MaxRetries: 1,
+		ShouldFailover: func(_ context.Context, out *schema.Message, err error) bool {
+			atomic.AddInt32(&shouldCalls, 1)
+			require.ErrorIs(t, err, streamErr)
+			require.NotNil(t, out)
+			require.Equal(t, "p1p2", out.Content)
+			return true
+		},
+		GetFailoverModel: func(_ context.Context, failoverCtx *FailoverContext[*schema.Message]) (model.BaseChatModel, []*schema.Message, error) {
+			require.Equal(t, uint(1), failoverCtx.FailoverAttempt)
+			return m2, nil, nil
+		},
+	}
+
+	wrapped := buildModelWrappers[*schema.Message](m1, &modelWrapperConfig{
+		failoverConfig: failoverCfg,
+	})
+
+	ctx := withTypedChatModelAgentExecCtx(context.Background(), &chatModelAgentExecCtx{
+		failoverLastSuccessModel: m1,
+	})
+	sr, err := wrapped.Stream(ctx, []*schema.Message{schema.UserMessage("hi")})
+	require.NoError(t, err)
+	msgs, err := drainMessageStream(sr)
+	require.NoError(t, err)
+	require.Len(t, msgs, 1)
+	require.Equal(t, "final", msgs[0].Content)
+	require.Equal(t, int32(1), atomic.LoadInt32(&m1Calls))
+	require.Equal(t, int32(1), atomic.LoadInt32(&m2Calls))
+	require.Equal(t, int32(1), atomic.LoadInt32(&shouldCalls))
+}
+
+func TestFailoverAcceptsAgenticAgent(t *testing.T) {
+	ctx := context.Background()
+
+	m := &mockAgenticModel{
+		generateFn: func(ctx context.Context, input []*schema.AgenticMessage, opts ...model.Option) (*schema.AgenticMessage, error) {
+			return agenticMsg("ok"), nil
+		},
+	}
+
+	fallbackModel := &mockAgenticModel{
+		generateFn: func(ctx context.Context, input []*schema.AgenticMessage, opts ...model.Option) (*schema.AgenticMessage, error) {
+			return agenticMsg("fallback"), nil
+		},
+	}
+
+	agent, err := NewTypedChatModelAgent(ctx, &TypedChatModelAgentConfig[*schema.AgenticMessage]{
+		Name:        "FailoverAgent",
+		Description: "Agent with failover config",
+		Model:       m,
+		ModelFailoverConfig: &ModelFailoverConfig[*schema.AgenticMessage]{
+			MaxRetries: 1,
+			ShouldFailover: func(ctx context.Context, outputMessage *schema.AgenticMessage, outputErr error) bool {
+				return true
+			},
+			GetFailoverModel: func(ctx context.Context, failoverCtx *FailoverContext[*schema.AgenticMessage]) (model.BaseModel[*schema.AgenticMessage], []*schema.AgenticMessage, error) {
+				return fallbackModel, nil, nil
+			},
+		},
+	})
+	require.NoError(t, err)
+	assert.NotNil(t, agent)
+}
+
+// approvalInfoSpan and approvalResultSpan are isolated copies for use in this
+// test file so we don't conflict with the prebuilt/integration_test.go types
+// (which live in a different package anyway).
+type approvalInfoSpan struct {
+	ToolName        string
+	ArgumentsInJSON string
+	ToolCallID      string
+}
+
+type approvalResultSpan struct {
+	Approved bool
+}
+
+func init() {
+	schema.Register[*approvalInfoSpan]()
+	schema.Register[*approvalResultSpan]()
+}
+
+// approvableSpanTool is an invokable tool that interrupts on first invocation
+// and runs to completion on resume after approval.
+type approvableSpanTool struct {
+	name string
+}
+
+func (t *approvableSpanTool) Info(_ context.Context) (*schema.ToolInfo, error) {
+	return &schema.ToolInfo{
+		Name: t.name,
+		Desc: "approvable span tool",
+		ParamsOneOf: schema.NewParamsOneOfByParams(map[string]*schema.ParameterInfo{
+			"input": {Type: schema.String, Desc: "input"},
+		}),
+	}, nil
+}
+
+func (t *approvableSpanTool) InvokableRun(ctx context.Context, argumentsInJSON string, _ ...tool.Option) (string, error) {
+	wasInterrupted, _, savedArgs := tool.GetInterruptState[string](ctx)
+	if !wasInterrupted {
+		return "", tool.StatefulInterrupt(ctx, &approvalInfoSpan{
+			ToolName:        t.name,
+			ArgumentsInJSON: argumentsInJSON,
+			ToolCallID:      compose.GetToolCallID(ctx),
+		}, argumentsInJSON)
+	}
+	isResumeTarget, hasData, data := tool.GetResumeContext[*approvalResultSpan](ctx)
+	if !isResumeTarget || !hasData {
+		return "", tool.StatefulInterrupt(ctx, &approvalInfoSpan{
+			ToolName:        t.name,
+			ArgumentsInJSON: savedArgs,
+			ToolCallID:      compose.GetToolCallID(ctx),
+		}, savedArgs)
+	}
+	if data.Approved {
+		return fmt.Sprintf("Tool '%s' executed with args: %s", t.name, savedArgs), nil
+	}
+	return fmt.Sprintf("Tool '%s' rejected", t.name), nil
+}
+
+// approvableStreamableSpanTool: streamable variant. Interrupts on first
+// invocation by returning a *core.InterruptSignal error before any stream
+// chunk is produced; runs to completion on resume.
+type approvableStreamableSpanTool struct {
+	name string
+}
+
+func (t *approvableStreamableSpanTool) Info(_ context.Context) (*schema.ToolInfo, error) {
+	return &schema.ToolInfo{
+		Name: t.name,
+		Desc: "approvable streamable span tool",
+		ParamsOneOf: schema.NewParamsOneOfByParams(map[string]*schema.ParameterInfo{
+			"input": {Type: schema.String, Desc: "input"},
+		}),
+	}, nil
+}
+
+func (t *approvableStreamableSpanTool) StreamableRun(ctx context.Context, argumentsInJSON string, _ ...tool.Option) (*schema.StreamReader[string], error) {
+	wasInterrupted, _, savedArgs := tool.GetInterruptState[string](ctx)
+	if !wasInterrupted {
+		return nil, tool.StatefulInterrupt(ctx, &approvalInfoSpan{
+			ToolName:        t.name,
+			ArgumentsInJSON: argumentsInJSON,
+			ToolCallID:      compose.GetToolCallID(ctx),
+		}, argumentsInJSON)
+	}
+	isResumeTarget, hasData, data := tool.GetResumeContext[*approvalResultSpan](ctx)
+	if !isResumeTarget || !hasData {
+		return nil, tool.StatefulInterrupt(ctx, &approvalInfoSpan{
+			ToolName:        t.name,
+			ArgumentsInJSON: savedArgs,
+			ToolCallID:      compose.GetToolCallID(ctx),
+		}, savedArgs)
+	}
+	if data.Approved {
+		return schema.StreamReaderFromArray([]string{
+			fmt.Sprintf("Tool '%s' streamed with args: %s", t.name, savedArgs),
+		}), nil
+	}
+	return schema.StreamReaderFromArray([]string{"rejected"}), nil
+}
+
+// alwaysErrorTool errors out hard (non-interrupt) on every invocation.
+type alwaysErrorTool struct {
+	name string
+}
+
+func (t *alwaysErrorTool) Info(_ context.Context) (*schema.ToolInfo, error) {
+	return &schema.ToolInfo{
+		Name: t.name,
+		Desc: "always errors",
+		ParamsOneOf: schema.NewParamsOneOfByParams(map[string]*schema.ParameterInfo{
+			"input": {Type: schema.String, Desc: "input"},
+		}),
+	}, nil
+}
+
+func (t *alwaysErrorTool) InvokableRun(_ context.Context, _ string, _ ...tool.Option) (string, error) {
+	return "", errors.New("hard tool failure")
+}
+
+// memCheckpointStore is a minimal in-memory CheckPointStore for these tests.
+type memCheckpointStore struct {
+	mu   sync.Mutex
+	data map[string][]byte
+}
+
+func newMemCheckpointStore() *memCheckpointStore {
+	return &memCheckpointStore{data: make(map[string][]byte)}
+}
+
+func (s *memCheckpointStore) Set(_ context.Context, key string, value []byte) error {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+	s.data[key] = value
+	return nil
+}
+
+func (s *memCheckpointStore) Get(_ context.Context, key string) ([]byte, bool, error) {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+	v, ok := s.data[key]
+	return v, ok, nil
+}
+
+// scriptedToolCallingModel is a controllable mock model: each call returns the
+// next scripted message.
+type scriptedToolCallingModel struct {
+	mu       sync.Mutex
+	messages []*schema.Message
+	pos      int
+}
+
+func (m *scriptedToolCallingModel) Generate(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.Message, error) {
+	m.mu.Lock()
+	defer m.mu.Unlock()
+	if m.pos >= len(m.messages) {
+		return schema.AssistantMessage("done", nil), nil
+	}
+	msg := m.messages[m.pos]
+	m.pos++
+	return msg, nil
+}
+
+func (m *scriptedToolCallingModel) Stream(ctx context.Context, input []*schema.Message, opts ...model.Option) (*schema.StreamReader[*schema.Message], error) {
+	msg, err := m.Generate(ctx, input, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return schema.StreamReaderFromArray([]*schema.Message{msg}), nil
+}
+
+func (m *scriptedToolCallingModel) WithTools(_ []*schema.ToolInfo) (model.ToolCallingChatModel, error) {
+	return m, nil
+}
+
+// drainAndCollectSpans collects tool span events emitted during iter draining.
+func drainAndCollectSpans(t *testing.T, iter *AsyncIterator[*AgentEvent]) (starts, ends []*SessionEvent[*schema.Message], interrupted bool) {
+	t.Helper()
+	for {
+		ev, ok := iter.Next()
+		if !ok {
+			break
+		}
+		if ev.Action != nil && ev.Action.Interrupted != nil {
+			interrupted = true
+		}
+		if ev.SessionEvent == nil || ev.SessionEvent.Span == nil {
+			continue
+		}
+		switch ev.SessionEvent.Kind {
+		case SessionEventSpanToolCallStart:
+			starts = append(starts, ev.SessionEvent)
+		case SessionEventSpanToolCallEnd:
+			ends = append(ends, ev.SessionEvent)
+		}
+	}
+	return
+}
+
+// setupApprovableSpanAgent constructs a ChatModelAgent with a scripted
+// tool-calling model and an in-memory checkpoint store, ready for span tests
+// that exercise interrupt/resume.
+func setupApprovableSpanAgent(t *testing.T, name string, tools []tool.BaseTool, scriptedAssistant []*schema.Message) (*TypedChatModelAgent[*schema.Message], *memCheckpointStore) {
+	t.Helper()
+	ctx := context.Background()
+	mdl := &scriptedToolCallingModel{messages: scriptedAssistant}
+	agent, err := NewChatModelAgent(ctx, &ChatModelAgentConfig{
+		Name:        name,
+		Description: "test",
+		Model:       mdl,
+		ToolsConfig: ToolsConfig{
+			ToolsNodeConfig: compose.ToolsNodeConfig{Tools: tools},
+		},
+	})
+	require.NoError(t, err)
+	store := newMemCheckpointStore()
+	return agent, store
+}
+
+func TestToolSpan_PermissionInterruptDefersEndSpan(t *testing.T) {
+	ctx := context.Background()
+	tl := &approvableSpanTool{name: "approve_me"}
+
+	scripted := []*schema.Message{
+		schema.AssistantMessage("calling", []schema.ToolCall{{ID: "call_1", Function: schema.FunctionCall{Name: tl.name, Arguments: `{"input":"x"}`}}}),
+	}
+	agent, store := setupApprovableSpanAgent(t, "agent1", []tool.BaseTool{tl}, scripted)
+	runner := NewRunner(ctx, RunnerConfig{Agent: agent, CheckPointStore: store})
+
+	checkpointID := "ckpt-1"
+	iter := runner.Run(ctx, []Message{schema.UserMessage("go")}, WithCheckPointID(checkpointID), WithTimelineEvents())
+	starts, ends, interrupted := drainAndCollectSpans(t, iter)
+
+	assert.True(t, interrupted, "expected an interrupt event from the approvable tool")
+	require.Len(t, starts, 1, "exactly one tool_call_start span should be emitted on the interrupted run")
+	assert.Empty(t, ends, "no tool_call_end span should be emitted on the interrupted run")
+	assert.Equal(t, "call_1", starts[0].Span.Tool.ToolUseID)
+	assert.NotEmpty(t, starts[0].Span.Tool.AssistantMessageEventID, "start span must carry assistant message event ID")
+	assert.NotEmpty(t, starts[0].Span.ParentSpanID, "start span must carry parent (model) span ID")
+}
+
+// runInterruptResumeAndCollectSpans drives a single-tool interrupt+approve
+// scenario and returns the start span emitted on the original run plus the
+// end span emitted on the resumed run. Used by both the resume happy-path
+// test and the dedicated "parent IDs survive resume" assertion.
+func runInterruptResumeAndCollectSpans(t *testing.T, agent *TypedChatModelAgent[*schema.Message], store *memCheckpointStore, checkpointID string) (startSpan, endSpan *SessionEvent[*schema.Message]) {
+	t.Helper()
+	ctx := context.Background()
+	runner := NewRunner(ctx, RunnerConfig{Agent: agent, CheckPointStore: store})
+	iter1 := runner.Run(ctx, []Message{schema.UserMessage("go")}, WithCheckPointID(checkpointID), WithTimelineEvents())
+
+	var (
+		starts1      []*SessionEvent[*schema.Message]
+		ends1        []*SessionEvent[*schema.Message]
+		interruptEvt *AgentEvent
+	)
+	for {
+		ev, ok := iter1.Next()
+		if !ok {
+			break
+		}
+		if ev.Action != nil && ev.Action.Interrupted != nil {
+			interruptEvt = ev
+		}
+		if ev.SessionEvent != nil && ev.SessionEvent.Span != nil {
+			switch ev.SessionEvent.Kind {
+			case SessionEventSpanToolCallStart:
+				starts1 = append(starts1, ev.SessionEvent)
+			case SessionEventSpanToolCallEnd:
+				ends1 = append(ends1, ev.SessionEvent)
+			}
+		}
+	}
+	require.NotNil(t, interruptEvt)
+	require.Len(t, starts1, 1)
+	require.Empty(t, ends1)
+
+	var toolInterruptID string
+	for _, ictx := range interruptEvt.Action.Interrupted.InterruptContexts {
+		if ictx.IsRootCause {
+			toolInterruptID = ictx.ID
+			break
+		}
+	}
+	require.NotEmpty(t, toolInterruptID)
+
+	resumeIter, err := runner.ResumeWithParams(ctx, checkpointID, &ResumeParams{
+		Targets: map[string]any{toolInterruptID: &approvalResultSpan{Approved: true}},
+	}, WithTimelineEvents())
+	require.NoError(t, err)
+	_, ends2, _ := drainAndCollectSpans(t, resumeIter)
+	require.Len(t, ends2, 1)
+	return starts1[0], ends2[0]
+}
+
+func TestToolSpan_PermissionResumeEmitsEndSpan(t *testing.T) {
+	tl := &approvableSpanTool{name: "approve_me"}
+	scripted := []*schema.Message{
+		schema.AssistantMessage("calling", []schema.ToolCall{{ID: "call_resume", Function: schema.FunctionCall{Name: tl.name, Arguments: `{"input":"x"}`}}}),
+		schema.AssistantMessage("done", nil),
+	}
+	agent, store := setupApprovableSpanAgent(t, "agent_resume", []tool.BaseTool{tl}, scripted)
+	startSpan, endSpan := runInterruptResumeAndCollectSpans(t, agent, store, "ckpt-resume")
+
+	assert.Equal(t, startSpan.Span.SpanID, endSpan.Span.SpanID, "end span must reuse the original SpanID")
+	assert.Equal(t, startSpan.EventID, endSpan.Span.Tool.ToolCallStartEventID, "end's ToolCallStartEventID must match start's EventID")
+	assert.Equal(t, "ok", endSpan.Span.Status)
+	assert.NotEmpty(t, endSpan.Span.Tool.ToolResultMessageEventID)
+}
+
+// TestToolSpan_ResumeUsesOriginalTurnParentIDs (plan §4.5.1 #3) verifies that
+// the resumed end span's ParentSpanID and AssistantMessageEventID match the
+// original turn's model span and assistant message — confirming the in-flight
+// span snapshot survived the checkpoint round-trip.
+func TestToolSpan_ResumeUsesOriginalTurnParentIDs(t *testing.T) {
+	tl := &approvableSpanTool{name: "approve_me"}
+	scripted := []*schema.Message{
+		schema.AssistantMessage("calling", []schema.ToolCall{{ID: "call_parents", Function: schema.FunctionCall{Name: tl.name, Arguments: `{"input":"x"}`}}}),
+		schema.AssistantMessage("done", nil),
+	}
+	agent, store := setupApprovableSpanAgent(t, "agent_parents", []tool.BaseTool{tl}, scripted)
+	startSpan, endSpan := runInterruptResumeAndCollectSpans(t, agent, store, "ckpt-parents")
+
+	require.NotEmpty(t, startSpan.Span.ParentSpanID, "start span carries a non-empty parent (model) span ID")
+	require.NotEmpty(t, startSpan.Span.Tool.AssistantMessageEventID, "start span carries a non-empty assistant message event ID")
+	assert.Equal(t, startSpan.Span.ParentSpanID, endSpan.Span.ParentSpanID, "ParentSpanID survives resume via the in-flight snapshot")
+	assert.Equal(t, startSpan.Span.Tool.AssistantMessageEventID, endSpan.Span.Tool.AssistantMessageEventID, "AssistantMessageEventID survives resume via the in-flight snapshot")
+}
+
+func TestToolSpan_HardErrorOnFirstRunStillEmitsEnd(t *testing.T) {
+	ctx := context.Background()
+	tl := &alwaysErrorTool{name: "boom"}
+	scripted := []*schema.Message{
+		schema.AssistantMessage("calling", []schema.ToolCall{{ID: "err_call", Function: schema.FunctionCall{Name: tl.name, Arguments: `{"input":"x"}`}}}),
+	}
+	agent, store := setupApprovableSpanAgent(t, "err_agent", []tool.BaseTool{tl}, scripted)
+	runner := NewRunner(ctx, RunnerConfig{Agent: agent, CheckPointStore: store})
+	iter := runner.Run(ctx, []Message{schema.UserMessage("go")}, WithCheckPointID("ckpt-err"), WithTimelineEvents())
+	starts, ends, _ := drainAndCollectSpans(t, iter)
+
+	require.Len(t, starts, 1)
+	require.Len(t, ends, 1)
+	assert.Equal(t, starts[0].Span.SpanID, ends[0].Span.SpanID, "end span shares SpanID with start span")
+	assert.Equal(t, "error", ends[0].Span.Status)
+}
+
+func TestToolSpan_StreamableInterruptDefersEnd(t *testing.T) {
+	ctx := context.Background()
+	tl := &approvableStreamableSpanTool{name: "stream_approve_me"}
+	scripted := []*schema.Message{
+		schema.AssistantMessage("calling", []schema.ToolCall{{ID: "stream_call", Function: schema.FunctionCall{Name: tl.name, Arguments: `{"input":"x"}`}}}),
+		schema.AssistantMessage("done", nil),
+	}
+	agent, store := setupApprovableSpanAgent(t, "stream_agent", []tool.BaseTool{tl}, scripted)
+	runner1 := NewRunner(ctx, RunnerConfig{Agent: agent, CheckPointStore: store})
+	checkpointID := "ckpt-stream"
+	iter1 := runner1.Run(ctx, []Message{schema.UserMessage("go")}, WithCheckPointID(checkpointID), WithTimelineEvents())
+	var interruptEvt *AgentEvent
+	starts1, ends1 := []*SessionEvent[*schema.Message]{}, []*SessionEvent[*schema.Message]{}
+	for {
+		ev, ok := iter1.Next()
+		if !ok {
+			break
+		}
+		if ev.Action != nil && ev.Action.Interrupted != nil {
+			interruptEvt = ev
+		}
+		if ev.SessionEvent != nil && ev.SessionEvent.Span != nil {
+			switch ev.SessionEvent.Kind {
+			case SessionEventSpanToolCallStart:
+				starts1 = append(starts1, ev.SessionEvent)
+			case SessionEventSpanToolCallEnd:
+				ends1 = append(ends1, ev.SessionEvent)
+			}
+		}
+	}
+	require.NotNil(t, interruptEvt)
+	require.Len(t, starts1, 1, "one start span on interrupted streamable run")
+	assert.Empty(t, ends1, "no end span on interrupted streamable run")
+	startSpanID := starts1[0].Span.SpanID
+
+	var toolInterruptID string
+	for _, ictx := range interruptEvt.Action.Interrupted.InterruptContexts {
+		if ictx.IsRootCause {
+			toolInterruptID = ictx.ID
+			break
+		}
+	}
+	require.NotEmpty(t, toolInterruptID)
+
+	resumeIter, err := runner1.ResumeWithParams(ctx, checkpointID, &ResumeParams{
+		Targets: map[string]any{toolInterruptID: &approvalResultSpan{Approved: true}},
+	}, WithTimelineEvents())
+	require.NoError(t, err)
+	starts2, ends2, _ := drainAndCollectSpans(t, resumeIter)
+	assert.Empty(t, starts2, "no new start span on streamable resume")
+	require.Len(t, ends2, 1, "one end span on streamable resume")
+	assert.Equal(t, startSpanID, ends2[0].Span.SpanID, "end span shares SpanID with start span across resume")
+	assert.Equal(t, "ok", ends2[0].Span.Status)
+}
+
+func TestTypedState_ToolSpansInFlightGobRoundTrip(t *testing.T) {
+	original := &typedState[*schema.Message]{
+		Messages:                       []*schema.Message{schema.UserMessage("hello")},
+		CurrentModelSpanID:             "model-span-1",
+		CurrentAssistantMessageEventID: "asst-event-1",
+		ToolSpansInFlight: map[string]*toolSpanInFlight{
+			"call_a": {
+				SpanID:                  "span-a",
+				StartEventID:            "start-event-a",
+				StartedAt:               time.Date(2026, 5, 26, 12, 0, 0, 0, time.UTC),
+				ParentSpanID:            "model-span-1",
+				AssistantMessageEventID: "asst-event-1",
+			},
+			"call_b": {
+				SpanID:                  "span-b",
+				StartEventID:            "start-event-b",
+				StartedAt:               time.Date(2026, 5, 26, 12, 0, 1, 0, time.UTC),
+				ParentSpanID:            "model-span-1",
+				AssistantMessageEventID: "asst-event-1",
+			},
+		},
+	}
+
+	var buf bytes.Buffer
+	require.NoError(t, gob.NewEncoder(&buf).Encode(original))
+
+	decoded := &typedState[*schema.Message]{}
+	require.NoError(t, gob.NewDecoder(&buf).Decode(decoded))
+
+	assert.Equal(t, original.CurrentModelSpanID, decoded.CurrentModelSpanID)
+	assert.Equal(t, original.CurrentAssistantMessageEventID, decoded.CurrentAssistantMessageEventID)
+	require.Len(t, decoded.ToolSpansInFlight, 2)
+	for k, v := range original.ToolSpansInFlight {
+		got, ok := decoded.ToolSpansInFlight[k]
+		require.Truef(t, ok, "missing key %q after gob round-trip", k)
+		assert.Equal(t, v.SpanID, got.SpanID)
+		assert.Equal(t, v.StartEventID, got.StartEventID)
+		assert.True(t, v.StartedAt.Equal(got.StartedAt), "StartedAt mismatch: %v vs %v", v.StartedAt, got.StartedAt)
+		assert.Equal(t, v.ParentSpanID, got.ParentSpanID)
+		assert.Equal(t, v.AssistantMessageEventID, got.AssistantMessageEventID)
+	}
+}
+
+// Sanity guard: ensure compose.IsInterruptRerunError import is preserved (used in wrappers).
+var _ = compose.IsInterruptRerunError
+
+// TestToolSpan_PermissionRejectEmitsEndSpan exercises the path where the tool
+// is interrupted, then on resume the user rejects (Approved=false). The tool
+// returns a rejection result rather than an error, so the end span carries
+// Status=ok with a populated ToolResultMessageEventID. Same SpanID across
+// the boundary.
+func TestToolSpan_PermissionRejectEmitsEndSpan(t *testing.T) {
+	ctx := context.Background()
+	tl := &approvableSpanTool{name: "reject_me"}
+
+	scripted := []*schema.Message{
+		schema.AssistantMessage("calling", []schema.ToolCall{{ID: "rej_call", Function: schema.FunctionCall{Name: tl.name, Arguments: `{"input":"x"}`}}}),
+		schema.AssistantMessage("done", nil),
+	}
+	agent, store := setupApprovableSpanAgent(t, "reject_agent", []tool.BaseTool{tl}, scripted)
+	checkpointID := "ckpt-reject"
+	runner := NewRunner(ctx, RunnerConfig{Agent: agent, CheckPointStore: store})
+	iter1 := runner.Run(ctx, []Message{schema.UserMessage("go")}, WithCheckPointID(checkpointID), WithTimelineEvents())
+
+	var (
+		starts1      []*SessionEvent[*schema.Message]
+		ends1        []*SessionEvent[*schema.Message]
+		interruptEvt *AgentEvent
+	)
+	for {
+		ev, ok := iter1.Next()
+		if !ok {
+			break
+		}
+		if ev.Action != nil && ev.Action.Interrupted != nil {
+			interruptEvt = ev
+		}
+		if ev.SessionEvent != nil && ev.SessionEvent.Span != nil {
+			switch ev.SessionEvent.Kind {
+			case SessionEventSpanToolCallStart:
+				starts1 = append(starts1, ev.SessionEvent)
+			case SessionEventSpanToolCallEnd:
+				ends1 = append(ends1, ev.SessionEvent)
+			}
+		}
+	}
+	require.NotNil(t, interruptEvt)
+	require.Len(t, starts1, 1)
+	require.Empty(t, ends1)
+
+	startSpanID := starts1[0].Span.SpanID
+
+	var toolInterruptID string
+	for _, ictx := range interruptEvt.Action.Interrupted.InterruptContexts {
+		if ictx.IsRootCause {
+			toolInterruptID = ictx.ID
+			break
+		}
+	}
+	require.NotEmpty(t, toolInterruptID)
+
+	resumeIter, err := runner.ResumeWithParams(ctx, checkpointID, &ResumeParams{
+		Targets: map[string]any{toolInterruptID: &approvalResultSpan{Approved: false}},
+	}, WithTimelineEvents())
+	require.NoError(t, err)
+	starts2, ends2, _ := drainAndCollectSpans(t, resumeIter)
+	assert.Empty(t, starts2)
+	require.Len(t, ends2, 1)
+	assert.Equal(t, startSpanID, ends2[0].Span.SpanID)
+	assert.Equal(t, "ok", ends2[0].Span.Status, "rejection produces a successful return (the deny content) — status is ok, not error")
+	assert.NotEmpty(t, ends2[0].Span.Tool.ToolResultMessageEventID)
+}
+
+// successOnlyTool runs to completion on the first invocation. Combined with
+// the absence of a permission middleware, it exercises the "non-interrupted
+// call" path where start and end both fire on the same run with status=ok.
+type successOnlyTool struct {
+	name   string
+	result string
+}
+
+func (t *successOnlyTool) Info(_ context.Context) (*schema.ToolInfo, error) {
+	return &schema.ToolInfo{
+		Name: t.name,
+		Desc: "always succeeds",
+		ParamsOneOf: schema.NewParamsOneOfByParams(map[string]*schema.ParameterInfo{
+			"input": {Type: schema.String, Desc: "input"},
+		}),
+	}, nil
+}
+
+func (t *successOnlyTool) InvokableRun(_ context.Context, _ string, _ ...tool.Option) (string, error) {
+	return t.result, nil
+}
+
+// TestToolSpan_NonInterruptedCallEmitsBothSpansOnSameRun verifies that a tool
+// that runs straight to success produces a tool_call_start + tool_call_end
+// pair on the same run, with the in-flight entry cleared at end emission.
+// (This corresponds to plan §4.5.1 #6 which used a "gate=deny" example —
+// the wire-shape behavior is identical: single run, single span pair, status
+// ok, populated ToolResultMessageEventID.)
+func TestToolSpan_NonInterruptedCallEmitsBothSpansOnSameRun(t *testing.T) {
+	ctx := context.Background()
+	tl := &successOnlyTool{name: "noninterrupted_tool", result: "ok"}
+	scripted := []*schema.Message{
+		schema.AssistantMessage("calling", []schema.ToolCall{{ID: "noninter_call", Function: schema.FunctionCall{Name: tl.name, Arguments: `{"input":"x"}`}}}),
+		schema.AssistantMessage("done", nil),
+	}
+	agent, store := setupApprovableSpanAgent(t, "noninter_agent", []tool.BaseTool{tl}, scripted)
+	runner := NewRunner(ctx, RunnerConfig{Agent: agent, CheckPointStore: store})
+	iter := runner.Run(ctx, []Message{schema.UserMessage("go")}, WithCheckPointID("ckpt-noninter"), WithTimelineEvents())
+	starts, ends, _ := drainAndCollectSpans(t, iter)
+
+	require.Len(t, starts, 1)
+	require.Len(t, ends, 1)
+	assert.Equal(t, starts[0].Span.SpanID, ends[0].Span.SpanID)
+	assert.Equal(t, "ok", ends[0].Span.Status)
+	assert.NotEmpty(t, ends[0].Span.Tool.ToolResultMessageEventID)
+}
+
+// TestToolSpan_ParallelInterruptResumesEmitMatchingEnds exercises the parallel
+// call scenario: two tool calls (A, B) emitted in a single assistant message,
+// both interrupting on first invocation. After the first run we should see
+// 2 starts and 0 ends. After resuming both with approval, we expect end spans
+// keyed to the matching SpanIDs (one per CallID).
+func TestToolSpan_ParallelInterruptResumesEmitMatchingEnds(t *testing.T) {
+	ctx := context.Background()
+	tl := &approvableSpanTool{name: "parallel_tool"}
+	scripted := []*schema.Message{
+		schema.AssistantMessage("calling 2", []schema.ToolCall{
+			{ID: "call_par_a", Function: schema.FunctionCall{Name: tl.name, Arguments: `{"input":"a"}`}},
+			{ID: "call_par_b", Function: schema.FunctionCall{Name: tl.name, Arguments: `{"input":"b"}`}},
+		}),
+		schema.AssistantMessage("done", nil),
+	}
+	agent, store := setupApprovableSpanAgent(t, "parallel_agent", []tool.BaseTool{tl}, scripted)
+	checkpointID := "ckpt-parallel"
+	runner := NewRunner(ctx, RunnerConfig{Agent: agent, CheckPointStore: store})
+	iter1 := runner.Run(ctx, []Message{schema.UserMessage("go")}, WithCheckPointID(checkpointID), WithTimelineEvents())
+
+	var (
+		starts1      []*SessionEvent[*schema.Message]
+		ends1        []*SessionEvent[*schema.Message]
+		interruptEvt *AgentEvent
+	)
+	for {
+		ev, ok := iter1.Next()
+		if !ok {
+			break
+		}
+		if ev.Action != nil && ev.Action.Interrupted != nil {
+			interruptEvt = ev
+		}
+		if ev.SessionEvent != nil && ev.SessionEvent.Span != nil {
+			switch ev.SessionEvent.Kind {
+			case SessionEventSpanToolCallStart:
+				starts1 = append(starts1, ev.SessionEvent)
+			case SessionEventSpanToolCallEnd:
+				ends1 = append(ends1, ev.SessionEvent)
+			}
+		}
+	}
+	require.NotNil(t, interruptEvt)
+	require.Len(t, starts1, 2, "expected one tool_call_start for each parallel call")
+	assert.Empty(t, ends1)
+
+	// Map CallID -> start SpanID for later assertions.
+	callIDToStartSpanID := map[string]string{}
+	for _, s := range starts1 {
+		callIDToStartSpanID[s.Span.Tool.ToolUseID] = s.Span.SpanID
+	}
+	require.Contains(t, callIDToStartSpanID, "call_par_a")
+	require.Contains(t, callIDToStartSpanID, "call_par_b")
+
+	// Collect interrupt IDs (root causes only).
+	var interruptIDs []string
+	for _, ictx := range interruptEvt.Action.Interrupted.InterruptContexts {
+		if ictx.IsRootCause {
+			interruptIDs = append(interruptIDs, ictx.ID)
+		}
+	}
+	require.Len(t, interruptIDs, 2)
+
+	// Approve both at once.
+	targets := map[string]any{}
+	for _, id := range interruptIDs {
+		targets[id] = &approvalResultSpan{Approved: true}
+	}
+	resumeIter, err := runner.ResumeWithParams(ctx, checkpointID, &ResumeParams{Targets: targets}, WithTimelineEvents())
+	require.NoError(t, err)
+	starts2, ends2, _ := drainAndCollectSpans(t, resumeIter)
+	assert.Empty(t, starts2, "no new starts on resume")
+	require.Len(t, ends2, 2, "two ends, one per parallel call")
+	for _, e := range ends2 {
+		expectedSpanID, ok := callIDToStartSpanID[e.Span.Tool.ToolUseID]
+		require.Truef(t, ok, "end span carries unknown CallID %q", e.Span.Tool.ToolUseID)
+		assert.Equal(t, expectedSpanID, e.Span.SpanID, "end span SpanID matches the start span for the same CallID")
+		assert.Equal(t, "ok", e.Span.Status)
+	}
+}
+
+func newFakeChatModel(
+	gen func(context.Context, []*schema.Message, ...model.Option) (*schema.Message, error),
+	stream func(context.Context, []*schema.Message, ...model.Option) (*schema.StreamReader[*schema.Message], error),
+) *fakeChatModel {
+	if gen == nil {
+		gen = func(context.Context, []*schema.Message, ...model.Option) (*schema.Message, error) {
+			return nil, errors.New("unused")
+		}
+	}
+	if stream == nil {
+		stream = func(context.Context, []*schema.Message, ...model.Option) (*schema.StreamReader[*schema.Message], error) {
+			return nil, errors.New("unused")
+		}
+	}
+	return &fakeChatModel{callbacksEnabled: true, generate: gen, stream: stream}
+}
+
+func TestRetryThenFailover(t *testing.T) {
+	t.Run("Generate_RetryExhaustedTriggersFailover", func(t *testing.T) {
+		modelErr := errors.New("model error")
+		var m1Calls int32
+		var m2Calls int32
+
+		m1 := newFakeChatModel(func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.Message, error) {
+			atomic.AddInt32(&m1Calls, 1)
+			return nil, modelErr
+		}, nil)
+		m2 := newFakeChatModel(func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.Message, error) {
+			atomic.AddInt32(&m2Calls, 1)
+			return schema.AssistantMessage("ok from m2", nil), nil
+		}, nil)
+
+		retryCfg := &ModelRetryConfig{
+			MaxRetries:  2,
+			IsRetryAble: func(_ context.Context, err error) bool { return true },
+			BackoffFunc: func(_ context.Context, _ int) time.Duration { return 0 },
+		}
+
+		failoverCfg := &ModelFailoverConfig[*schema.Message]{
+			MaxRetries: 1,
+			ShouldFailover: func(_ context.Context, _ *schema.Message, err error) bool {
+				return err != nil
+			},
+			GetFailoverModel: func(_ context.Context, fc *FailoverContext[*schema.Message]) (model.BaseChatModel, []*schema.Message, error) {
+				require.NotNil(t, fc.LastErr)
+				return m2, nil, nil
+			},
+		}
+
+		wrapped := buildModelWrappers[*schema.Message](m1, &modelWrapperConfig{
+			retryConfig:    retryCfg,
+			failoverConfig: failoverCfg,
+		})
+
+		ctx := withTypedChatModelAgentExecCtx(context.Background(), &chatModelAgentExecCtx{
+			failoverLastSuccessModel: m1,
+		})
+		msg, err := wrapped.Generate(ctx, []*schema.Message{schema.UserMessage("hi")})
+		require.NoError(t, err)
+		require.Equal(t, "ok from m2", msg.Content)
+
+		// m1: 1 (lastSuccess) + 2 retries = 3 calls on lastSuccess attempt,
+		// then failover to m2 which also goes through retry wrapper: 1 call succeeds.
+		require.Equal(t, int32(3), atomic.LoadInt32(&m1Calls))
+		require.Equal(t, int32(1), atomic.LoadInt32(&m2Calls))
+	})
+
+	t.Run("Generate_AllExhausted", func(t *testing.T) {
+		modelErr := errors.New("always fails")
+		var m1Calls int32
+		var m2Calls int32
+
+		m1 := newFakeChatModel(func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.Message, error) {
+			atomic.AddInt32(&m1Calls, 1)
+			return nil, modelErr
+		}, nil)
+		m2 := newFakeChatModel(func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.Message, error) {
+			atomic.AddInt32(&m2Calls, 1)
+			return nil, modelErr
+		}, nil)
+
+		retryCfg := &ModelRetryConfig{
+			MaxRetries:  1,
+			IsRetryAble: func(_ context.Context, err error) bool { return true },
+			BackoffFunc: func(_ context.Context, _ int) time.Duration { return 0 },
+		}
+
+		failoverCfg := &ModelFailoverConfig[*schema.Message]{
+			MaxRetries: 1,
+			ShouldFailover: func(_ context.Context, _ *schema.Message, err error) bool {
+				return err != nil
+			},
+			GetFailoverModel: func(_ context.Context, _ *FailoverContext[*schema.Message]) (model.BaseChatModel, []*schema.Message, error) {
+				return m2, nil, nil
+			},
+		}
+
+		wrapped := buildModelWrappers[*schema.Message](m1, &modelWrapperConfig{
+			retryConfig:    retryCfg,
+			failoverConfig: failoverCfg,
+		})
+
+		ctx := withTypedChatModelAgentExecCtx(context.Background(), &chatModelAgentExecCtx{
+			failoverLastSuccessModel: m1,
+		})
+		_, err := wrapped.Generate(ctx, []*schema.Message{schema.UserMessage("hi")})
+		require.Error(t, err)
+
+		// Should be RetryExhaustedError from m2's retry wrapper
+		var retryErr *RetryExhaustedError
+		require.True(t, errors.As(err, &retryErr))
+
+		// m1: 1 initial + 1 retry = 2 calls
+		require.Equal(t, int32(2), atomic.LoadInt32(&m1Calls))
+		// m2: 1 initial + 1 retry = 2 calls
+		require.Equal(t, int32(2), atomic.LoadInt32(&m2Calls))
+	})
+
+	t.Run("Generate_RetrySucceedsNoFailover", func(t *testing.T) {
+		var m1Calls int32
+		var failoverCalled int32
+
+		m1 := newFakeChatModel(func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.Message, error) {
+			n := atomic.AddInt32(&m1Calls, 1)
+			if n == 1 {
+				return nil, errors.New("transient error")
+			}
+			return schema.AssistantMessage("ok on retry", nil), nil
+		}, nil)
+
+		retryCfg := &ModelRetryConfig{
+			MaxRetries:  2,
+			IsRetryAble: func(_ context.Context, err error) bool { return true },
+			BackoffFunc: func(_ context.Context, _ int) time.Duration { return 0 },
+		}
+
+		failoverCfg := &ModelFailoverConfig[*schema.Message]{
+			MaxRetries: 1,
+			ShouldFailover: func(_ context.Context, _ *schema.Message, err error) bool {
+				atomic.AddInt32(&failoverCalled, 1)
+				return true
+			},
+			GetFailoverModel: func(_ context.Context, _ *FailoverContext[*schema.Message]) (model.BaseChatModel, []*schema.Message, error) {
+				t.Fatal("GetFailoverModel should not be called when retry succeeds")
+				return nil, nil, nil
+			},
+		}
+
+		wrapped := buildModelWrappers[*schema.Message](m1, &modelWrapperConfig{
+			retryConfig:    retryCfg,
+			failoverConfig: failoverCfg,
+		})
+
+		ctx := withTypedChatModelAgentExecCtx(context.Background(), &chatModelAgentExecCtx{
+			failoverLastSuccessModel: m1,
+		})
+		msg, err := wrapped.Generate(ctx, []*schema.Message{schema.UserMessage("hi")})
+		require.NoError(t, err)
+		require.Equal(t, "ok on retry", msg.Content)
+
+		// 2 calls: first fails, second succeeds via retry
+		require.Equal(t, int32(2), atomic.LoadInt32(&m1Calls))
+		// ShouldFailover should never be called
+		require.Equal(t, int32(0), atomic.LoadInt32(&failoverCalled))
+	})
+
+	t.Run("Generate_NonRetryableErrorTriggersFailover", func(t *testing.T) {
+		nonRetryableErr := errors.New("non-retryable")
+		var m1Calls int32
+		var m2Calls int32
+
+		m1 := newFakeChatModel(func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.Message, error) {
+			atomic.AddInt32(&m1Calls, 1)
+			return nil, nonRetryableErr
+		}, nil)
+		m2 := newFakeChatModel(func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.Message, error) {
+			atomic.AddInt32(&m2Calls, 1)
+			return schema.AssistantMessage("ok from m2", nil), nil
+		}, nil)
+
+		retryCfg := &ModelRetryConfig{
+			MaxRetries: 3,
+			IsRetryAble: func(_ context.Context, err error) bool {
+				// Only non-retryable errors
+				return !errors.Is(err, nonRetryableErr)
+			},
+			BackoffFunc: func(_ context.Context, _ int) time.Duration { return 0 },
+		}
+
+		failoverCfg := &ModelFailoverConfig[*schema.Message]{
+			MaxRetries: 1,
+			ShouldFailover: func(_ context.Context, _ *schema.Message, err error) bool {
+				return err != nil
+			},
+			GetFailoverModel: func(_ context.Context, _ *FailoverContext[*schema.Message]) (model.BaseChatModel, []*schema.Message, error) {
+				return m2, nil, nil
+			},
+		}
+
+		wrapped := buildModelWrappers[*schema.Message](m1, &modelWrapperConfig{
+			retryConfig:    retryCfg,
+			failoverConfig: failoverCfg,
+		})
+
+		ctx := withTypedChatModelAgentExecCtx(context.Background(), &chatModelAgentExecCtx{
+			failoverLastSuccessModel: m1,
+		})
+		msg, err := wrapped.Generate(ctx, []*schema.Message{schema.UserMessage("hi")})
+		require.NoError(t, err)
+		require.Equal(t, "ok from m2", msg.Content)
+
+		// m1 called only once — non-retryable error skips retry
+		require.Equal(t, int32(1), atomic.LoadInt32(&m1Calls))
+		require.Equal(t, int32(1), atomic.LoadInt32(&m2Calls))
+	})
+
+	t.Run("Stream_RetryExhaustedTriggersFailover", func(t *testing.T) {
+		streamErr := errors.New("stream mid error")
+		var m1Calls int32
+		var m2Calls int32
+
+		m1 := newFakeChatModel(nil, func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.StreamReader[*schema.Message], error) {
+			atomic.AddInt32(&m1Calls, 1)
+			return streamWithMidError([]*schema.Message{
+				schema.AssistantMessage("partial", nil),
+			}, streamErr), nil
+		})
+		m2 := newFakeChatModel(nil, func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.StreamReader[*schema.Message], error) {
+			atomic.AddInt32(&m2Calls, 1)
+			return schema.StreamReaderFromArray([]*schema.Message{schema.AssistantMessage("ok from m2", nil)}), nil
+		})
+
+		retryCfg := &ModelRetryConfig{
+			MaxRetries:  1,
+			IsRetryAble: func(_ context.Context, err error) bool { return true },
+			BackoffFunc: func(_ context.Context, _ int) time.Duration { return 0 },
+		}
+
+		failoverCfg := &ModelFailoverConfig[*schema.Message]{
+			MaxRetries: 1,
+			ShouldFailover: func(_ context.Context, _ *schema.Message, err error) bool {
+				return err != nil
+			},
+			GetFailoverModel: func(_ context.Context, fc *FailoverContext[*schema.Message]) (model.BaseChatModel, []*schema.Message, error) {
+				require.NotNil(t, fc.LastErr)
+				return m2, nil, nil
+			},
+		}
+
+		wrapped := buildModelWrappers[*schema.Message](m1, &modelWrapperConfig{
+			retryConfig:    retryCfg,
+			failoverConfig: failoverCfg,
+		})
+
+		ctx := withTypedChatModelAgentExecCtx(context.Background(), &chatModelAgentExecCtx{
+			failoverLastSuccessModel: m1,
+		})
+		sr, err := wrapped.Stream(ctx, []*schema.Message{schema.UserMessage("hi")})
+		require.NoError(t, err)
+		msgs, err := drainMessageStream(sr)
+		require.NoError(t, err)
+		require.Len(t, msgs, 1)
+		require.Equal(t, "ok from m2", msgs[0].Content)
+
+		// m1: 1 initial + 1 retry = 2 calls on lastSuccess attempt
+		require.Equal(t, int32(2), atomic.LoadInt32(&m1Calls))
+		require.Equal(t, int32(1), atomic.LoadInt32(&m2Calls))
+	})
+
+	t.Run("Stream_AllExhausted", func(t *testing.T) {
+		streamErr := errors.New("always fails mid-stream")
+		var m1Calls int32
+		var m2Calls int32
+
+		m1 := newFakeChatModel(nil, func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.StreamReader[*schema.Message], error) {
+			atomic.AddInt32(&m1Calls, 1)
+			return streamWithMidError([]*schema.Message{
+				schema.AssistantMessage("p", nil),
+			}, streamErr), nil
+		})
+		m2 := newFakeChatModel(nil, func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.StreamReader[*schema.Message], error) {
+			atomic.AddInt32(&m2Calls, 1)
+			return streamWithMidError([]*schema.Message{
+				schema.AssistantMessage("p", nil),
+			}, streamErr), nil
+		})
+
+		retryCfg := &ModelRetryConfig{
+			MaxRetries:  1,
+			IsRetryAble: func(_ context.Context, err error) bool { return true },
+			BackoffFunc: func(_ context.Context, _ int) time.Duration { return 0 },
+		}
+
+		failoverCfg := &ModelFailoverConfig[*schema.Message]{
+			MaxRetries: 1,
+			ShouldFailover: func(_ context.Context, _ *schema.Message, err error) bool {
+				return err != nil
+			},
+			GetFailoverModel: func(_ context.Context, _ *FailoverContext[*schema.Message]) (model.BaseChatModel, []*schema.Message, error) {
+				return m2, nil, nil
+			},
+		}
+
+		wrapped := buildModelWrappers[*schema.Message](m1, &modelWrapperConfig{
+			retryConfig:    retryCfg,
+			failoverConfig: failoverCfg,
+		})
+
+		ctx := withTypedChatModelAgentExecCtx(context.Background(), &chatModelAgentExecCtx{
+			failoverLastSuccessModel: m1,
+		})
+		_, err := wrapped.Stream(ctx, []*schema.Message{schema.UserMessage("hi")})
+		require.Error(t, err)
+
+		var retryErr *RetryExhaustedError
+		require.True(t, errors.As(err, &retryErr))
+
+		// m1: 1 initial + 1 retry = 2 calls
+		require.Equal(t, int32(2), atomic.LoadInt32(&m1Calls))
+		// m2: 1 initial + 1 retry = 2 calls
+		require.Equal(t, int32(2), atomic.LoadInt32(&m2Calls))
+	})
+
+	t.Run("ShouldRetry_Stream_TriggersFailover", func(t *testing.T) {
+		var m1Calls int32
+		var m2Calls int32
+
+		m1 := newFakeChatModel(nil, func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.StreamReader[*schema.Message], error) {
+			atomic.AddInt32(&m1Calls, 1)
+			return schema.StreamReaderFromArray([]*schema.Message{schema.AssistantMessage("bad from m1", nil)}), nil
+		})
+		m2 := newFakeChatModel(nil, func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.StreamReader[*schema.Message], error) {
+			atomic.AddInt32(&m2Calls, 1)
+			return schema.StreamReaderFromArray([]*schema.Message{schema.AssistantMessage("good from m2", nil)}), nil
+		})
+
+		retryCfg := &ModelRetryConfig{
+			MaxRetries: 1,
+			ShouldRetry: func(_ context.Context, retryCtx *RetryContext) *RetryDecision {
+				if retryCtx.OutputMessage != nil && retryCtx.OutputMessage.Content == "bad from m1" {
+					return &RetryDecision{Retry: true}
+				}
+				return &RetryDecision{Retry: false}
+			},
+			BackoffFunc: func(_ context.Context, _ int) time.Duration { return 0 },
+		}
+
+		failoverCfg := &ModelFailoverConfig[*schema.Message]{
+			MaxRetries: 1,
+			ShouldFailover: func(_ context.Context, _ *schema.Message, err error) bool {
+				return err != nil
+			},
+			GetFailoverModel: func(_ context.Context, _ *FailoverContext[*schema.Message]) (model.BaseChatModel, []*schema.Message, error) {
+				return m2, nil, nil
+			},
+		}
+
+		wrapped := buildModelWrappers[*schema.Message](m1, &modelWrapperConfig{
+			retryConfig:    retryCfg,
+			failoverConfig: failoverCfg,
+		})
+
+		ctx := withTypedChatModelAgentExecCtx(context.Background(), &chatModelAgentExecCtx{
+			failoverLastSuccessModel: m1,
+		})
+		sr, err := wrapped.Stream(ctx, []*schema.Message{schema.UserMessage("hi")})
+		require.NoError(t, err)
+		msgs, err := drainMessageStream(sr)
+		require.NoError(t, err)
+		require.Len(t, msgs, 1)
+		require.Equal(t, "good from m2", msgs[0].Content)
+		require.Equal(t, int32(2), atomic.LoadInt32(&m1Calls))
+		require.Equal(t, int32(1), atomic.LoadInt32(&m2Calls))
+	})
+
+	t.Run("ShouldRetry_Generate_TriggersFailover", func(t *testing.T) {
+		var m1Calls int32
+		var m2Calls int32
+
+		m1 := newFakeChatModel(func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.Message, error) {
+			atomic.AddInt32(&m1Calls, 1)
+			return schema.AssistantMessage("bad from m1", nil), nil
+		}, nil)
+		m2 := newFakeChatModel(func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.Message, error) {
+			atomic.AddInt32(&m2Calls, 1)
+			return schema.AssistantMessage("good from m2", nil), nil
+		}, nil)
+
+		retryCfg := &ModelRetryConfig{
+			MaxRetries: 1,
+			ShouldRetry: func(_ context.Context, retryCtx *RetryContext) *RetryDecision {
+				if retryCtx.OutputMessage != nil && retryCtx.OutputMessage.Content == "bad from m1" {
+					return &RetryDecision{Retry: true}
+				}
+				return &RetryDecision{Retry: false}
+			},
+			BackoffFunc: func(_ context.Context, _ int) time.Duration { return 0 },
+		}
+
+		failoverCfg := &ModelFailoverConfig[*schema.Message]{
+			MaxRetries: 1,
+			ShouldFailover: func(_ context.Context, _ *schema.Message, err error) bool {
+				return err != nil
+			},
+			GetFailoverModel: func(_ context.Context, _ *FailoverContext[*schema.Message]) (model.BaseChatModel, []*schema.Message, error) {
+				return m2, nil, nil
+			},
+		}
+
+		wrapped := buildModelWrappers[*schema.Message](m1, &modelWrapperConfig{
+			retryConfig:    retryCfg,
+			failoverConfig: failoverCfg,
+		})
+
+		ctx := withTypedChatModelAgentExecCtx(context.Background(), &chatModelAgentExecCtx{
+			failoverLastSuccessModel: m1,
+		})
+		msg, err := wrapped.Generate(ctx, []*schema.Message{schema.UserMessage("hi")})
+		require.NoError(t, err)
+		require.Equal(t, "good from m2", msg.Content)
+		require.Equal(t, int32(2), atomic.LoadInt32(&m1Calls))
+		require.Equal(t, int32(1), atomic.LoadInt32(&m2Calls))
+	})
+
+	t.Run("Stream_GetFailoverModelReturnsNilModel", func(t *testing.T) {
+		streamErr := errors.New("m1 always fails")
+		var m1Calls int32
+
+		m1 := newFakeChatModel(nil, func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.StreamReader[*schema.Message], error) {
+			atomic.AddInt32(&m1Calls, 1)
+			return nil, streamErr
+		})
+
+		retryCfg := &ModelRetryConfig{
+			MaxRetries:  0,
+			IsRetryAble: func(_ context.Context, err error) bool { return false },
+			BackoffFunc: func(_ context.Context, _ int) time.Duration { return 0 },
+		}
+
+		failoverCfg := &ModelFailoverConfig[*schema.Message]{
+			MaxRetries: 1,
+			ShouldFailover: func(_ context.Context, _ *schema.Message, err error) bool {
+				return err != nil
+			},
+			GetFailoverModel: func(_ context.Context, _ *FailoverContext[*schema.Message]) (model.BaseChatModel, []*schema.Message, error) {
+				return nil, nil, nil
+			},
+		}
+
+		wrapped := buildModelWrappers[*schema.Message](m1, &modelWrapperConfig{
+			retryConfig:    retryCfg,
+			failoverConfig: failoverCfg,
+		})
+
+		ctx := withTypedChatModelAgentExecCtx(context.Background(), &chatModelAgentExecCtx{
+			failoverLastSuccessModel: m1,
+		})
+		_, err := wrapped.Stream(ctx, []*schema.Message{schema.UserMessage("hi")})
+		require.Error(t, err)
+		require.Contains(t, err.Error(), "returned nil model at attempt")
+		require.Equal(t, int32(1), atomic.LoadInt32(&m1Calls))
+	})
+
+	t.Run("Stream_ContextCanceledDuringFailover", func(t *testing.T) {
+		streamErr := errors.New("m1 fails")
+		var m1Calls int32
+		var failoverModelCalled int32
+
+		m1 := newFakeChatModel(nil, func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.StreamReader[*schema.Message], error) {
+			atomic.AddInt32(&m1Calls, 1)
+			return nil, streamErr
+		})
+
+		ctx, cancel := context.WithCancel(context.Background())
+
+		retryCfg := &ModelRetryConfig{
+			MaxRetries:  0,
+			IsRetryAble: func(_ context.Context, err error) bool { return false },
+			BackoffFunc: func(_ context.Context, _ int) time.Duration { return 0 },
+		}
+
+		failoverCfg := &ModelFailoverConfig[*schema.Message]{
+			MaxRetries: 3,
+			ShouldFailover: func(_ context.Context, _ *schema.Message, err error) bool {
+				cancel()
+				return err != nil
+			},
+			GetFailoverModel: func(_ context.Context, _ *FailoverContext[*schema.Message]) (model.BaseChatModel, []*schema.Message, error) {
+				atomic.AddInt32(&failoverModelCalled, 1)
+				return nil, nil, nil
+			},
+		}
+
+		wrapped := buildModelWrappers[*schema.Message](m1, &modelWrapperConfig{
+			retryConfig:    retryCfg,
+			failoverConfig: failoverCfg,
+		})
+
+		ctx = withTypedChatModelAgentExecCtx(ctx, &chatModelAgentExecCtx{
+			failoverLastSuccessModel: m1,
+		})
+		_, err := wrapped.Stream(ctx, []*schema.Message{schema.UserMessage("hi")})
+		require.Error(t, err)
+		require.ErrorIs(t, err, context.Canceled)
+		require.Equal(t, int32(1), atomic.LoadInt32(&m1Calls))
+		require.Equal(t, int32(0), atomic.LoadInt32(&failoverModelCalled))
+	})
+}
+
+func TestErrStreamCanceled_Failover(t *testing.T) {
+	t.Run("Stream_NeverFailedOver", func(t *testing.T) {
+		var m1Calls int32
+		var failoverCalled int32
+
+		m1 := newFakeChatModel(nil, func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.StreamReader[*schema.Message], error) {
+			atomic.AddInt32(&m1Calls, 1)
+			return streamWithMidError([]*schema.Message{
+				schema.AssistantMessage("partial", nil),
+			}, ErrStreamCanceled), nil
+		})
+
+		failoverCfg := &ModelFailoverConfig[*schema.Message]{
+			MaxRetries: 2,
+			ShouldFailover: func(_ context.Context, _ *schema.Message, err error) bool {
+				atomic.AddInt32(&failoverCalled, 1)
+				return true
+			},
+			GetFailoverModel: func(_ context.Context, _ *FailoverContext[*schema.Message]) (model.BaseChatModel, []*schema.Message, error) {
+				t.Fatal("GetFailoverModel should not be called for ErrStreamCanceled")
+				return nil, nil, nil
+			},
+		}
+
+		wrapped := buildModelWrappers[*schema.Message](m1, &modelWrapperConfig{
+			failoverConfig: failoverCfg,
+		})
+
+		ctx := withTypedChatModelAgentExecCtx(context.Background(), &chatModelAgentExecCtx{
+			failoverLastSuccessModel: m1,
+		})
+		_, err := wrapped.Stream(ctx, []*schema.Message{schema.UserMessage("hi")})
+		require.Error(t, err)
+		require.True(t, errors.Is(err, ErrStreamCanceled))
+		require.Equal(t, int32(1), atomic.LoadInt32(&m1Calls))
+		require.Equal(t, int32(0), atomic.LoadInt32(&failoverCalled))
+	})
+
+	t.Run("Generate_NeverFailedOver", func(t *testing.T) {
+		var m1Calls int32
+		var failoverCalled int32
+
+		m1 := newFakeChatModel(func(_ context.Context, _ []*schema.Message, _ ...model.Option) (*schema.Message, error) {
+			atomic.AddInt32(&m1Calls, 1)
+			return nil, ErrStreamCanceled
+		}, nil)
+
+		failoverCfg := &ModelFailoverConfig[*schema.Message]{
+			MaxRetries: 2,
+			ShouldFailover: func(_ context.Context, _ *schema.Message, err error) bool {
+				atomic.AddInt32(&failoverCalled, 1)
+				return true
+			},
+			GetFailoverModel: func(_ context.Context, _ *FailoverContext[*schema.Message]) (model.BaseChatModel, []*schema.Message, error) {
+				t.Fatal("GetFailoverModel should not be called for ErrStreamCanceled")
+				return nil, nil, nil
+			},
+		}
+
+		wrapped := buildModelWrappers[*schema.Message](m1, &modelWrapperConfig{
+			failoverConfig: failoverCfg,
+		})
+
+		ctx := withTypedChatModelAgentExecCtx(context.Background(), &chatModelAgentExecCtx{
+			failoverLastSuccessModel: m1,
+		})
+		_, err := wrapped.Generate(ctx, []*schema.Message{schema.UserMessage("hi")})
+		require.Error(t, err)
+		require.True(t, errors.Is(err, ErrStreamCanceled))
+		require.Equal(t, int32(1), atomic.LoadInt32(&m1Calls))
+		require.Equal(t, int32(0), atomic.LoadInt32(&failoverCalled))
+	})
 }
